refactor: remove AgentGraph, extract shared types, add resizable graph panel
